moved health check info into its own component
