+++ golang/dashboard/src/components/Relay-Info.vue
@@ -28,7 +28,7 @@
           </v-layout>
 
           <v-flex class="relay__info__info" row>
-            <p>Height: {{ height }}</p>
+            <p>Height: {{ currentBlock.height }}</p>
             <Click-To-Copy :copy-value="height"/>
           </v-flex>
 
@@ -41,11 +41,12 @@
             <Click-To-Copy :copy-value="currentBlock.hash"/>
           </v-flex>
 
-          <v-flex class="relay__info__info" row>
+          <!-- <v-flex class="relay__info__info" row>
+            TODO: Fix verifiedAt
             <p v-if="verifiedAt === null">Unverified</p>
             <p v-else-if="verifiedAt < 1">Verified: Less than 1 minute ago</p>
             <p v-else>Verified: {{ verifiedAt }} minute<span v-if="verifiedAt > 1">s</span> ago</p>
-          </v-flex>
+          </v-flex> -->
         </v-layout>
 
         <v-divider/>
@@ -90,115 +91,41 @@
       </div>
     </v-card>
 
-    <v-card class="relay__updates">
-      <v-layout>
-        <p class="relay__updates__title">Relay Health Check:</p>
-        <p v-if="lastCommsRelay === null">Not completed</p>
-        <p v-else-if="lastCommsRelay < 1">Less than 1 minute ago</p>
-        <p v-else>{{ lastCommsRelay }} minute<span v-if="lastCommsRelay > 1">s</span> ago</p>
-      </v-layout>
-      <v-layout>
-        <p class="relay__updates__title">Source:</p>
-        <p>{{ source }}</p>
-      </v-layout>
-      <v-layout>
-        <p class="relay__updates__title">Source Health Check:</p>
-        <p v-if="lastCommsExternal === null">Not completed</p>
-        <p v-else-if="lastCommsExternal < 1">Less than 1 minute ago</p>
-        <p v-else>{{ lastCommsExternal }} minute<span v-if="lastCommsExternal > 1">s</span> ago</p>
-      </v-layout>
-      <v-layout>
-        <p class="relay__updates__title">Source Block Changed:</p>
-        <p v-if="verifiedAt === null">Unknown</p>
-        <p v-else-if="verifiedAt < 1">Less than 1 minute ago</p>
-        <p v-else>{{ verifiedAt }} minute<span v-if="verifiedAt > 1">s</span> ago</p>
-      </v-layout>
-      <v-layout>
-        <p class="relay__updates__title">Source Height:</p>
-        <p>{{ height || 'Unknown' }}</p>
-      </v-layout>
-    </v-card>
-
-    <!-- <v-card>
-      <v-btn @click="getBKD">Get BKD</v-btn>
-      <v-btn @click="getLCA">Get LCA</v-btn>
-      <v-btn @click="getCurrentHeight">Get Current Height</v-btn>
-    </v-card> -->
+    <Relay-Health-Check />
   </v-container>
 </template>
 
 <script>
 import { mapState } from 'vuex'
-import { getMinsAgo } from '@/utils/utils'
 
 export default {
-  name: 'relay',
+  name: 'Relay',
 
   components: {
+    RelayHealthCheck: () => import(/* webpackChunkName: 'Relay-Health-Check' */ './Relay-Health-Check'),
     ClickToCopy: () => import(/* webpackChunkName: 'Click-To-Copy' */ './Click-To-Copy'),
     NetType: () => import(/* webpackChunkName: 'Net-Type' */ './Net-Type')
   },
 
   data: () => ({
     windowWidth: Number,
-    verifiedAt: null,
-    lastCommsExternal: null,
-    lastCommsRelay: null
+    verifiedAt: null
   }),
+
+  computed: {
+    ...mapState({
+      currentBlock: state => state.info.currentBlock,
+      relay: state => state.info.relay
+    })
+  },
 
   mounted () {
     this.onResize()
-
-    // Calculate minutes for health check
-    this.healthCheckMins()
-
-    // Updates every minute
-    setInterval(() => {
-      this.healthCheckMins()
-    }, 60000)
   },
 
   methods: {
     onResize () {
       this.windowWidth = window.innerWidth
-    },
-
-    healthCheckMins () {
-      this.verifiedAt = this.currentBlock.verifiedAt ? getMinsAgo(this.currentBlock.verifiedAt) : null
-      this.lastCommsExternal = this.lastComms.external ? getMinsAgo(this.lastComms.external) : null
-      this.lastCommsRelay = this.lastComms.relay ? getMinsAgo(this.lastComms.relay) : null
-    }
-
-    // getBKD () {
-    //   this.$store.dispatch('relay/getBKD')
-    // },
-
-    // getLCA () {
-    //   this.$store.dispatch('relay/getLCA')
-    // },
-
-    // getCurrentHeight () {
-    //   this.$store.dispatch('relay/verifyHeight', this.currentBlock.hash)
-    // }
-  },
-
-  computed: {
-    ...mapState({
-      lastComms: state => state.info.lastComms,
-      currentBlock: state => state.info.currentBlock,
-      height: state => state.info.currentBlock.height,
-      relay: state => state.info.relay,
-      source: state => state.info.source
-    })
-  },
-
-  watch: {
-    lastComms: {
-      handler: function () {
-        console.log('Updated info')
-        this.healthCheckMins()
-      },
-      deep: true
     }
   },
 
@@ -240,23 +167,9 @@ export default {
   margin-left: 0px;
 }
 
-.relay__updates {
-  margin-top: 50px;
-  padding: 20px;
-  max-width: 500px;
-}
-
-.relay__updates__title {
-  width: 200px;
-  font-weight: 900;
-}
-
 @media (max-width: 800px) {
   .relay {
     padding: 0 20px;
-  }
-  .relay__updates {
-    max-width: none;
   }
 }
 </style>
